The enemy is assigned with a RaycastShooter that will shoot a projectile whenever an object moves in front of it.
