+++ RiverGame/Assets/Scenes/SampleScene.unity
@@ -312,8 +312,9 @@ GameObject:
   m_Component:
   - component: {fileID: 284768965}
   - component: {fileID: 284768964}
-  - component: {fileID: 284768963}
   - component: {fileID: 284768962}
+  - component: {fileID: 284768963}
+  - component: {fileID: 284768966}
   m_Layer: 0
   m_Name: Enemy
   m_TagString: Untagged
@@ -387,6 +388,17 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 1
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &284768966
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 0}
+  m_GameObject: {fileID: 284768961}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 96ffb665583d3c448a2ca325a6e66e48, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1 &301374822
 GameObject:
   m_ObjectHideFlags: 0
