+++ RiverGame/Assets/Scripts/Enemy/EnemyController.cs
@@ -5,14 +5,6 @@ using UnityEngine;
 public class EnemyController : MonoBehaviour
 {
 	[SerializeField] private int health;
-	[SerializeField] private GameObject projectile;
-
-	private float time;
-
-	public void Start()
-	{
-		time = Time.time;
-	}
 
 	public void Update()
 	{
@@ -20,18 +12,5 @@ public class EnemyController : MonoBehaviour
 		{
 			Destroy(gameObject);
 		}
-
-		if (time + 3f < Time.time)
-		{
-			time = Time.time;
-			Shoot();
-		}
-	}
-
-	public void Shoot()
-	{
-		Vector3 pos = new Vector3(transform.position.x, transform.position.y + 1, transform.position.z + 2f);
-		GameObject proj = Instantiate(projectile, pos, Quaternion.identity, transform.parent);
-		proj.GetComponent<Rigidbody>().velocity = new Vector3(10, 0, 0);
 	}
 }
