+++ RiverGame/Assets/Scripts/Shooting.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: fab8797979e818149bd37a6e945dfd02
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
