+++ RiverGame/Assets/Scripts/Shooting/RayShooter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 96ffb665583d3c448a2ca325a6e66e48
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
